🚧 1DH19J cloud: keep hotspot checks green
