+++ .agentplane/tasks/202605091226-1DH19J/README.md
@@ -19,9 +19,9 @@ plan_approval:
   note: null
 verification:
   state: "ok"
-  updated_at: "2026-05-09T12:26:38.539Z"
+  updated_at: "2026-05-09T12:29:50.051Z"
   updated_by: "CODER"
-  note: "Cloud pull Node fetch timeout fix verified with focused tests, typecheck, lint, bootstrap, and live pull."
+  note: "Cloud pull fetch timeout fix verified after hotspot-safe test split."
 commit: null
 comments:
   -
@@ -41,8 +41,14 @@ events:
     author: "CODER"
     state: "ok"
     note: "Cloud pull Node fetch timeout fix verified with focused tests, typecheck, lint, bootstrap, and live pull."
+  -
+    type: "verify"
+    at: "2026-05-09T12:29:50.051Z"
+    author: "CODER"
+    state: "ok"
+    note: "Cloud pull fetch timeout fix verified after hotspot-safe test split."
 doc_version: 3
-doc_updated_at: "2026-05-09T12:26:38.545Z"
+doc_updated_at: "2026-05-09T12:29:50.061Z"
 doc_updated_by: "CODER"
 description: "Normalize Node address-selection attempt timeout for cloud fetch so backend sync pull can refresh large cloud projections instead of failing with UND_ERR_CONNECT_TIMEOUT."
 sections:
@@ -81,6 +87,24 @@ sections:
     - route_changed: no
     - safe_command: agentplane blueprint snapshot 202605091226-1DH19J
     
+    ### 2026-05-09T12:29:50.051Z — VERIFY — ok
+    
+    By: CODER
+    
+    Note: Cloud pull fetch timeout fix verified after hotspot-safe test split.
+    
+    VerifyStepsRef: doc_version=3, doc_updated_at=2026-05-09T12:26:38.545Z, excerpt_hash=sha256:0c911ba57bbda86e6b1d4b2c31f39ff10ccc1febf923fdb7f66dbb574080a0d7
+    
+    Details:
+    
+    BlueprintSnapshotRef:
+    - state: current
+    - path: /Users/densmirnov/Github/agentplane/.agentplane/tasks/202605091226-1DH19J/blueprint/resolved-snapshot.json
+    - old_digest: b66b9eb8b8de1712464aaa28bafd9de0f4a1e7e1f3ff6ad737b29b426d975cf4
+    - current_digest: b66b9eb8b8de1712464aaa28bafd9de0f4a1e7e1f3ff6ad737b29b426d975cf4
+    - route_changed: no
+    - safe_command: agentplane blueprint snapshot 202605091226-1DH19J
+    
     <!-- END VERIFICATION RESULTS -->
   Rollback Plan: |-
     - Revert task-related commit(s).
@@ -89,5 +113,9 @@ sections:
     - Observation: Command: bun test packages/agentplane/src/backends/task-backend.cloud.test.ts | Result: pass | Evidence: 25 tests passed, including Node default 10000ms timeout normalization regression. Scope: cloud backend sync behavior.\nCommand: bun run typecheck | Result: pass | Evidence: tsc -b completed. Scope: workspace TypeScript contracts.\nCommand: bun run lint:core | Result: pass | Evidence: eslint completed. Scope: repo lint contract.\nCommand: bun run framework:dev:bootstrap && ap backend sync cloud --direction pull --conflict=diff --yes | Result: pass | Evidence: repo-local runtime rebuilt; live pull returned changed=0 ignored_remote_only=0 conflicts=0. Scope: live cloud projection refresh.
       Impact: Node fetch no longer keeps the unsafe 10000ms address-selection attempt timeout that produced UND_ERR_CONNECT_TIMEOUT against sync.agentplane.cloud.
       Resolution: Normalize cloud fetch address-selection attempt timeout to 1000ms and cover both too-low and Node-default cases.
+    
+    - Observation: Command: bun test packages/agentplane/src/backends/task-backend.cloud.test.ts packages/agentplane/src/backends/task-backend/cloud-backend-utils.test.ts | Result: pass | Evidence: 25 tests passed across cloud backend and cloud backend utility suites. Scope: sync behavior and address-selection timeout normalization.\nCommand: bun run hotspots:check | Result: pass | Evidence: cloud-backend.ts is 598 LoC; oversized test baseline remains 10 entries. Scope: hotspot guard.\nCommand: bun run typecheck | Result: pass | Evidence: tsc -b completed. Scope: workspace TypeScript contracts.\nCommand: bun run lint:core | Result: pass | Evidence: eslint completed. Scope: repo lint contract.\nCommand: bun run framework:dev:bootstrap && ap backend sync cloud --direction pull --conflict=diff --yes | Result: pass | Evidence: live pull returned changed=0 ignored_remote_only=0 conflicts=0. Scope: live cloud projection refresh.
+      Impact: Node fetch no longer keeps the unsafe 10000ms address-selection attempt timeout that produced UND_ERR_CONNECT_TIMEOUT; hotspot guard stays green without reintroducing allow-lists.
+      Resolution: Normalize cloud fetch address-selection attempt timeout to 1000ms and move focused utility tests out of the oversized cloud backend suite.
 id_source: "generated"
 ---
